fail on gobble([ notANode ])
